Dependency Inject the CountdownPublisher
Resulted in using a struct for the CountdownPublisher arguments, as we are injecting the function that creates the CountdownPublisher, not the instance itself,
as we don't know the instance parameters until run time. This requires passing the CountdownPublisher as a function, which (due to Swift issues) means that
we can't use labelled arguments (when passing a function).